made some new classes for organization , specifically when it comes to dif types of users
