+++ spring-boot/src/main/java/com/example/springboot/Guest.java
@@ -63,7 +63,7 @@ public class Guest implements User {
                 return "success";
             }
         }
-        return "failure";
+        return removedRoom;
     }
 
     //returns the String that was removed from the csv file (commas included)
@@ -122,54 +122,8 @@ public class Guest implements User {
         }
     }
 
-    public String addAvailableRoom(String newRoom){
-        InputStream is = this.getClass().getResourceAsStream("/RoomsAvailable.csv");
-
-        List<String> lines = new ArrayList<>();
-        try (BufferedReader reader = new BufferedReader(new InputStreamReader(is))) {
-            String line;
-            while ((line = reader.readLine()) != null) {
-                lines.add(line);
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-            return "Failed to read in the currently available rooms";
-        }
-
-        // Add the new line containing the new reservation
-        if(! lines.contains(newRoom)) {
-            lines.add(newRoom);
-        }
-        else{
-            return "Room is already there";
-        }
-
-        FileWriter fw;
-        try {
-            fw = new FileWriter("RoomsTaken.csv");
-        }
-        catch(IOException x){
-            x.printStackTrace();
-            return "Could not write to RoomsAvailable database";
-        }
-
-        // Write the updated content back to the CSV file
-        try (BufferedWriter writer = new BufferedWriter(fw))
-        {
-            for (String line : lines) {
-                writer.write(line);
-                writer.newLine();
-            }
-        }
-        catch (IOException e) {
-            e.printStackTrace();
-            return "Could not write to RoomsAvailable database";
-        }
-
-        return "success";    }
-
     //takes a csv formatted line and puts it into the RoomsTaken.csv
-    //returns either "success" or a fail message depending on result
+    //returns either success" or a fail message depending on result
     private String addReservedRoom(String newRoom){
         InputStream is = this.getClass().getResourceAsStream("/RoomsTaken.csv");
