+++ spring-boot/src/main/java/com/example/springboot/UserType.java
@@ -1,0 +1,5 @@
+package com.example.springboot;
+
+public enum UserType {
+
+}
