+++ spring-boot/src/main/resources/Users.csv
@@ -1,0 +1,2 @@
+username,password,userType
+catherinearling,catePass
